[develop] Clean up(Remove Delegate since I'm currently not using)
