+++ src/main/java/com/abn/recipeapi_v1/controllers/RecipeIngredientsApiController.java
@@ -1,31 +1,22 @@
 package com.abn.recipeapi_v1.controllers;
 
 import com.abn.recipeapi_v1.RecipeIngredientsApi;
-import com.abn.recipeapi_v1.RecipeIngredientsApiDelegate;
 import com.abn.recipeapi_v1.model.RecipeDTO;
 import com.abn.recipeapi_v1.services.RecipeIngredientDAOService;
-import lombok.Getter;
-import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import java.util.Optional;
 import java.util.UUID;
 
 @Controller
 @RequestMapping("${openapi.recipeIngredients.base-path:}")
 public class RecipeIngredientsApiController implements RecipeIngredientsApi {
 
-	@Getter
-    private final RecipeIngredientsApiDelegate delegate;
     private final RecipeIngredientDAOService service;
 
-    public RecipeIngredientsApiController(
-            @Autowired(required = false) RecipeIngredientsApiDelegate delegate,
-            RecipeIngredientDAOService service
-    ) {
-        this.delegate = Optional.ofNullable(delegate).orElse(new RecipeIngredientsApiDelegate() {});
+    public RecipeIngredientsApiController(RecipeIngredientDAOService service) {
         this.service = service;
     }
